+++ changelog.md
@@ -14,6 +14,7 @@
 - Fixed
   - Mod download notification url can't show more than 3 lines anymore.
   - Creating a new mod profile now uses your current mods.
+  - TriOS update notification shows download size before downloading. 
 - Changed
   - "Launch" button now reacts to clicks a bit faster and is more clear that the game's running.